+++ mmocr/apis/inferencers/base_inferencer.py
@@ -1,12 +1,9 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-import os.path as osp
 from typing import Dict, List, Optional, Sequence, Tuple, Union
 
-import mmcv
 import numpy as np
 import torch
 from mmengine import Config, InstanceData
-from mmengine.dataset import Compose
 from mmengine.runner import load_checkpoint
 
 from mmocr.registry import MODELS, VISUALIZERS
@@ -42,13 +39,7 @@ class BaseInferencer:
             Defaults to False.
     """
 
-    func_kwargs = dict(
-        preprocess=[],
-        forward=[],
-        visualize=[
-            'show', 'wait_time', 'draw_pred', 'pred_score_thr', 'img_out_dir'
-        ],
-        postprocess=['print_result', 'pred_out_file'])
+    func_kwargs = dict(preprocess=[], forward=[], visualize=[], postprocess=[])
     func_order = dict(preprocess=0, forward=1, visualize=2, postprocess=3)
 
     def __init__(self,
@@ -56,10 +47,6 @@ class BaseInferencer:
                  ckpt: Optional[str],
                  device: Optional[str] = None,
                  **kwargs) -> None:
-        # A global counter tracking the number of images processed, for
-        # naming of the output images
-        self.num_visualized_imgs = 0
-
         # Load config to cfg
         if isinstance(config, str):
             cfg = Config.fromfile(config)
@@ -91,38 +78,7 @@ class BaseInferencer:
 
     def _init_pipeline(self, cfg: ConfigType) -> None:
         """Initialize the test pipeline."""
-        pipeline_cfg = cfg.test_dataloader.dataset.pipeline
-
-        # For inference, the key of ``instances`` is not used.
-        if 'meta_keys' in pipeline_cfg[-1]:
-            pipeline_cfg[-1]['meta_keys'] = tuple(
-                meta_key for meta_key in pipeline_cfg[-1]['meta_keys']
-                if meta_key != 'instances')
-
-        # Loading annotations is also not applicable
-        idx = self._get_transform_idx(pipeline_cfg, 'LoadOCRAnnotations')
-        if idx != -1:
-            del pipeline_cfg[idx]
-
-        self.file_pipeline = Compose(pipeline_cfg)
-
-        load_img_idx = self._get_transform_idx(pipeline_cfg,
-                                               'LoadImageFromFile')
-        if load_img_idx == -1:
-            raise ValueError(
-                'LoadImageFromFile is not found in the test pipeline')
-        pipeline_cfg[load_img_idx] = dict(type='mmdet.LoadImageFromNDArray')
-        self.ndarray_pipeline = Compose(pipeline_cfg)
-
-    def _get_transform_idx(self, pipeline_cfg: ConfigType, name: str) -> int:
-        """Returns the index of the transform in a pipeline.
-
-        If the transform is not found, returns -1.
-        """
-        for i, transform in enumerate(pipeline_cfg):
-            if transform['type'] == name:
-                return i
-        return -1
+        raise NotImplementedError
 
     def _init_visualizer(self, cfg: ConfigType) -> None:
         """Initialize visualizers."""
@@ -153,22 +109,7 @@ class BaseInferencer:
 
         return results
 
-    def preprocess(self, inputs: InputsType) -> List[Dict]:
-        """Process the inputs into a model-feedable format."""
-        results = []
-        for single_input in inputs:
-            if isinstance(single_input, str):
-                data_ = dict(img_path=single_input)
-                results.append(self.file_pipeline(data_))
-            elif isinstance(single_input, np.ndarray):
-                data_ = dict(img=single_input)
-                results.append(self.ndarray_pipeline(data_))
-            else:
-                raise ValueError(
-                    f'Unsupported input type: {type(single_input)}')
-        return results
-
-    def __call__(self, user_inputs: InputsType,
+    def __call__(self, inputs: InputsType,
                  **kwargs) -> Union[Dict, List[Dict]]:
         """Call the inferencer.
 
@@ -176,10 +117,6 @@ class BaseInferencer:
             user_inputs: Inputs for the inferencer.
             kwargs: Keyword arguments for the inferencer.
         """
-
-        # Detect if user_inputs are in a batch
-        is_batch = isinstance(user_inputs, (list, tuple))
-        inputs = user_inputs if is_batch else [user_inputs]
 
         params = self.dispatch_kwargs(**kwargs)
         preprocess_kwargs = self.base_params[0].copy()
@@ -194,9 +131,12 @@ class BaseInferencer:
         data = self.preprocess(inputs, **preprocess_kwargs)
         preds = self.forward(data, **forward_kwargs)
         imgs = self.visualize(inputs, preds, **visualize_kwargs)
-        results = self.postprocess(
-            preds, imgs, is_batch=is_batch, **postprocess_kwargs)
+        results = self.postprocess(preds, imgs, **postprocess_kwargs)
         return results
+
+    def preprocess(self, inputs: InputsType) -> List[Dict]:
+        """Process the inputs into a model-feedable format."""
+        raise NotImplementedError
 
     def forward(self, inputs: InputsType) -> PredType:
         with torch.no_grad():
@@ -224,53 +164,13 @@ class BaseInferencer:
                 Defaults to 0.3.
             img_out_dir (str): Output directory of images. Defaults to ''.
         """
-        if self.visualizer is None or not show and img_out_dir == '':
-            return None
+        raise NotImplementedError
 
-        if getattr(self, 'visualizer') is None:
-            raise ValueError('Visualization needs the "visualizer" term'
-                             'defined in the config, but got None.')
-
-        results = []
-
-        for single_input, pred in zip(inputs, preds):
-            if isinstance(single_input, str):
-                img = mmcv.imread(single_input)
-                img_name = osp.basename(single_input)
-            elif isinstance(single_input, np.ndarray):
-                img = single_input.copy()
-                img_num = str(self.num_visualized_imgs).zfill(8)
-                img_name = f'{img_num}.jpg'
-            else:
-                raise ValueError('Unsupported input type: '
-                                 f'{type(single_input)}')
-
-            out_file = osp.join(img_out_dir, img_name) if img_out_dir != '' \
-                else None
-
-            self.visualizer.add_datasample(
-                img_name,
-                img,
-                pred,
-                show=show,
-                wait_time=wait_time,
-                draw_gt=False,
-                draw_pred=draw_pred,
-                pred_score_thr=pred_score_thr,
-                out_file=out_file,
-            )
-            results.append(img)
-            self.num_visualized_imgs += 1
-
-        return results
-
-    def postprocess(self,
-                    preds: PredType,
-                    imgs: Optional[List[np.ndarray]] = None,
-                    is_batch: bool = False,
-                    print_result: bool = False,
-                    pred_out_file: str = ''
-                    ) -> Union[ResType, Tuple[ResType, np.ndarray]]:
+    def postprocess(
+        self,
+        preds: PredType,
+        imgs: Optional[List[np.ndarray]] = None,
+    ) -> Union[ResType, Tuple[ResType, np.ndarray]]:
         """Postprocess predictions.
 
         Args:
@@ -286,27 +186,5 @@ class BaseInferencer:
 
         Returns:
             TODO
-        """
-        results = []
-        for pred in preds:
-            result = self._pred2dict(pred)
-            results.append(result)
-        if not is_batch:
-            results = results[0]
-        if print_result:
-            print(results)
-        # Add img to the results after printing
-        if pred_out_file != '':
-            mmcv.dump(results, pred_out_file)
-        if imgs is None:
-            return results
-        return results, imgs
-
-    def _pred2dict(self, data_sample: InstanceData) -> Dict:
-        """Extract elements necessary to represent a prediction into a
-        dictionary.
-
-        It's better to contain only basic data elements such as strings and
-        numbers in order to guarantee it's json-serializable.
         """
         raise NotImplementedError
